LayoutTests/imported/w3c/web-platform-tests/uievents/click/click_event_target_siblings-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Click targeting when targets of down and up are sibling elements

This test verifies that click event always goes to the first common ancestor of down and up event targets.
Press down the primary button on red div and move to blue box and release.
Press down the primary button on button b1 and move to button b2 and release.
Press down the primary button on input i1 and move to input i2 and release.
Press down the primary button on link 1 and move to link 2 and release.
Click done.
b1  b2

link1
link2
Done

Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN Click targets the nearest common ancestor
